Added Entity and EntityType.
Player, Wall, and Bullet were all changed to reflect the new system. Each one implements Entity which requires them to return their EntityType.

This allows the object to be added to the Body user data. This can later be retrieved or used in situations like collision.

Collision can call getEntityType() to know when certain things collide and propogate events to the game loop. Right now this is done in a crude way by simply adding bullet/wall collisions to a set of bullets to remove.

We use a set because the Box2D world step can have a bullet collide multiple times, and removing this twice from the world is a terrible idea (crashing).